FIX(usePrint): move to hook file
